+++ cpp/order/Steinhardt.cc
@@ -4,6 +4,7 @@
 #include "Steinhardt.h"
 #include "NeighborComputeFunctional.h"
 #include "utils.h"
+#include <vector>
 
 /*! \file Steinhardt.cc
     \brief Computes variants of Steinhardt order parameters.
@@ -12,47 +13,52 @@
 namespace freud { namespace order {
 
 // Calculating Ylm using fsph module
-void Steinhardt::computeYlm(const float theta, const float phi, std::vector<std::complex<float>>& Ylm) const
+void Steinhardt::computeYlm(fsph::PointSPHEvaluator<float>& sph_eval, const float theta, const float phi,
+                            YlmsType& Ylms) const
 {
-    if (Ylm.size() != m_num_ms)
-    {
-        Ylm.resize(m_num_ms);
-    }
-
-    fsph::PointSPHEvaluator<float> sph_eval(m_l);
-
-    unsigned int m_index(0);
     sph_eval.compute(theta, phi);
 
-    for (typename fsph::PointSPHEvaluator<float>::iterator iter(sph_eval.begin_l(m_l, 0, true));
-         iter != sph_eval.end(); ++iter)
+    for (size_t i = 0; i < m_ls.size(); ++i)
     {
-        // Manually add the Condon-Shortley phase, (-1)^m, to positive odd m
-        float phase = 1;
-        if (m_index <= m_l && m_index % 2 == 1)
-        {
-            phase = -1;
-        }
+        auto& Ylm = Ylms[i];
+        const auto l = m_ls[i];
+        size_t m_index(0);
 
-        Ylm[m_index] = phase * (*iter);
-        ++m_index;
+        const auto end_iter(sph_eval.begin_l(l + 1, 0, true));
+        for (auto iter(sph_eval.begin_l(l, 0, true)); iter != end_iter; ++iter, ++m_index)
+        {
+            // Manually add the Condon-Shortley phase, (-1)^m, to positive odd m
+            const float phase = (m_index <= l && m_index % 2 == 1) ? -1 : 1;
+            Ylm[m_index] = phase * (*iter);
+        }
     }
 }
 
 void Steinhardt::reallocateArrays(unsigned int Np)
 {
     m_Np = Np;
-    m_qlmi.prepare({Np, m_num_ms});
-    m_qlm.prepare(m_num_ms);
-    m_qli.prepare(Np);
+
+    const auto num_ls = m_ls.size();
+
+    m_qli.prepare({Np, num_ls});
     if (m_average)
     {
-        m_qlmiAve.prepare({Np, m_num_ms});
-        m_qliAve.prepare(Np);
+        m_qliAve.prepare({Np, num_ls});
     }
     if (m_wl)
     {
-        m_wli.prepare(Np);
+        m_wli.prepare({Np, num_ls});
+    }
+
+    for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
+    {
+        const auto num_ms = m_num_ms[l_index];
+        m_qlmi[l_index].prepare({Np, num_ms});
+        m_qlm[l_index].prepare(num_ms);
+        if (m_average)
+        {
+            m_qlmiAve[l_index].prepare({Np, num_ms});
+        }
     }
 }
 
@@ -71,7 +77,10 @@ void Steinhardt::compute(const freud::locality::NeighborList* nlist,
     }
 
     // Reduce qlm
-    m_qlm_local.reduceInto(m_qlm);
+    for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
+    {
+        m_qlm_local[l_index].reduceInto(m_qlm[l_index]);
+    }
 
     if (m_wl)
     {
@@ -90,15 +99,35 @@ void Steinhardt::compute(const freud::locality::NeighborList* nlist,
 void Steinhardt::baseCompute(const freud::locality::NeighborList* nlist,
                              const freud::locality::NeighborQuery* points, freud::locality::QueryArgs qargs)
 {
-    const auto normalizationfactor = float(4.0 * M_PI / m_num_ms);
+    std::vector<float> normalizationfactor(m_ls.size());
+    for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
+    {
+        normalizationfactor[l_index] = float(4.0 * M_PI / m_num_ms[l_index]);
+    }
     // For consistency, this reset is done here regardless of whether the array
     // is populated in baseCompute or computeAve.
-    m_qlm_local.reset();
+    for (auto& qlm_local : m_qlm_local)
+    {
+        qlm_local.reset();
+    }
+
     freud::locality::loopOverNeighborsIterator(
         points, points->getPoints(), m_Np, qargs, nlist,
         [=](size_t i, const std::shared_ptr<freud::locality::NeighborPerPointIterator>& ppiter) {
             float total_weight(0);
             const vec3<float> ref((*points)[i]);
+            // Construct PointSPHEvaluator outside loop since the construction is costly.
+            auto max_l = *std::max_element(m_ls.begin(), m_ls.end());
+            fsph::PointSPHEvaluator<float> sph_eval(max_l);
+
+            // Alocate and instantiate this array before looping over particles to prevent N instantiations
+            // and N * m_l.size() allocations.
+            YlmsType Ylms(m_ls.size());
+            for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
+            {
+                Ylms[l_index].resize(m_num_ms[l_index]);
+            }
+
             for (freud::locality::NeighborBond nb = ppiter->next(); !ppiter->end(); nb = ppiter->next())
             {
                 const vec3<float> delta = points->getBox().wrap((*points)[nb.point_idx] - ref);
@@ -107,7 +136,7 @@ void Steinhardt::baseCompute(const freud::locality::NeighborList* nlist,
                 // phi is usually in range 0..2Pi, but
                 // it only appears in Ylm as exp(im\phi),
                 // so range -Pi..Pi will give same results.
-                float phi = std::atan2(delta.y, delta.x); // -Pi..Pi
+                const float phi = std::atan2(delta.y, delta.x); // -Pi..Pi
 
                 // This value must be clamped in cases where the particles are
                 // aligned along z, otherwise due to floating point error we
@@ -122,32 +151,51 @@ void Steinhardt::baseCompute(const freud::locality::NeighborList* nlist,
                     theta = 0;
                 }
 
-                std::vector<std::complex<float>> Ylm(m_num_ms);
-                computeYlm(theta, phi, Ylm); // Fill up Ylm
+                computeYlm(sph_eval, theta, phi, Ylms); // Fill up Ylm
 
-                for (unsigned int k = 0; k < m_num_ms; ++k)
+                for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
                 {
-                    m_qlmi({static_cast<unsigned int>(i), k}) += weight * Ylm[k];
+                    auto& qlmi = m_qlmi[l_index];
+                    const auto& Ylm = Ylms[l_index];
+                    // Get the initial index and iterate using ++ for faster iteration
+                    // Profiling showed using operator() to slow the code significantly.
+                    const size_t index = qlmi.getIndex({i, 0});
+                    for (size_t k = 0; k < m_num_ms[l_index]; ++k)
+                    {
+                        qlmi[index + k] += weight * Ylm[k];
+                    }
                 }
+                // Accumulate weight for normalization
                 total_weight += weight;
             } // End loop going over neighbor bonds
 
             // Normalize!
-            for (unsigned int k = 0; k < m_num_ms; ++k)
+            const size_t qli_i_start = m_qli.getIndex({i, 0});
+            for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
             {
-                // Cache the index for efficiency.
-                const unsigned int index = m_qlmi.getIndex({static_cast<unsigned int>(i), k});
-                m_qlmi[index] /= total_weight;
-                // Add the norm, which is the (complex) squared magnitude
-                m_qli[i] += norm(m_qlmi[index]);
-                // This array gets populated by computeAve in the averaging case.
-                if (!m_average)
+                // get l specific vectors/arrays
+                auto& qlmi = m_qlmi[l_index];
+                auto& qlm_local = m_qlm_local[l_index];
+                const size_t first_qlmi_index = qlmi.getIndex({i, 0});
+                const size_t qli_index = qli_i_start + l_index;
+
+                for (size_t k = 0; k < m_num_ms[l_index]; ++k)
                 {
-                    m_qlm_local.local()[k] += m_qlmi[index] / float(m_Np);
+                    // Cache the index for efficiency.
+                    const size_t qlmi_index = first_qlmi_index + k;
+
+                    qlmi[qlmi_index] /= total_weight;
+                    // Add the norm, which is the (complex) squared magnitude
+                    m_qli[qli_index] += norm(qlmi[qlmi_index]);
+                    // This array gets populated by computeAve in the averaging case.
+                    if (!m_average)
+                    {
+                        qlm_local.local()[k] += qlmi[qlmi_index] / float(m_Np);
+                    }
                 }
+                m_qli[qli_index] *= normalizationfactor[l_index];
+                m_qli[qli_index] = std::sqrt(m_qli[qli_index]);
             }
-            m_qli[i] *= normalizationfactor;
-            m_qli[i] = std::sqrt(m_qli[i]);
         });
 }
 
@@ -160,7 +208,11 @@ void Steinhardt::computeAve(const freud::locality::NeighborList* nlist,
         iter = points->query(points->getPoints(), points->getNPoints(), qargs);
     }
 
-    const auto normalizationfactor = float(4.0 * M_PI / m_num_ms);
+    std::vector<float> normalizationfactor(m_ls.size());
+    for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
+    {
+        normalizationfactor[l_index] = static_cast<float>(4.0 * M_PI / m_num_ms[l_index]);
+    }
 
     freud::locality::loopOverNeighborsIterator(
         points, points->getPoints(), m_Np, qargs, nlist,
@@ -184,80 +236,118 @@ void Steinhardt::computeAve(const freud::locality::NeighborList* nlist,
                 for (freud::locality::NeighborBond nb2 = ns_neighbors_iter->next(); !ns_neighbors_iter->end();
                      nb2 = ns_neighbors_iter->next())
                 {
-                    for (unsigned int k = 0; k < m_num_ms; ++k)
+                    for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
                     {
-                        // Adding all the qlm of the neighbors. We use the
-                        // vector function signature for indexing into the
-                        // arrays for speed.
-                        m_qlmiAve({static_cast<unsigned int>(i), k}) += m_qlmi({nb2.point_idx, k});
+                        auto& qlmiAve = m_qlmiAve[l_index];
+                        auto& qlmi = m_qlmi[l_index];
+                        const auto ave_index = qlmiAve.getIndex({i, 0});
+                        const auto nb2_index = qlmi.getIndex({nb2.point_idx, 0});
+                        for (size_t k = 0; k < m_num_ms[l_index]; ++k)
+                        {
+                            // Adding all the qlm of the neighbors. We use the
+                            // vector function signature for indexing into the
+                            // arrays for speed.
+                            qlmiAve[ave_index + k] += qlmi[nb2_index + k];
+                        }
                     }
                     neighborcount++;
                 } // End loop over particle neighbor's bonds
             }     // End loop over particle's bonds
 
             // Normalize!
-            for (unsigned int k = 0; k < m_num_ms; ++k)
+
+            const size_t qliAve_i_start = m_qliAve.getIndex({i, 0});
+            for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
             {
-                // Cache the index for efficiency.
-                const unsigned int index = m_qlmiAve.getIndex({static_cast<unsigned int>(i), k});
-                // Adding the qlm of the particle i itself
-                m_qlmiAve[index] += m_qlmi[index];
-                m_qlmiAve[index] /= static_cast<float>(neighborcount);
-                m_qlm_local.local()[k] += m_qlmiAve[index] / float(m_Np);
-                // Add the norm, which is the complex squared magnitude
-                m_qliAve[i] += norm(m_qlmiAve[index]);
+                auto& qlmiAve = m_qlmiAve[l_index];
+                auto& qlmi = m_qlmi[l_index];
+                auto& qlm_local = m_qlm_local[l_index];
+                const size_t first_qlmi_index = qlmiAve.getIndex({i, 0});
+                const size_t qliAve_index = qliAve_i_start + l_index;
+
+                for (size_t k = 0; k < m_num_ms[l_index]; ++k)
+                {
+                    // Cache the index for efficiency.
+                    const size_t qlmi_index = first_qlmi_index + k;
+                    // Adding the qlm of the particle i itself
+                    qlmiAve[qlmi_index] += qlmi[qlmi_index];
+                    qlmiAve[qlmi_index] /= static_cast<float>(neighborcount);
+                    qlm_local.local()[k] += qlmiAve[qlmi_index] / float(m_Np);
+                    // Add the norm, which is the complex squared magnitude
+                    m_qliAve[qliAve_index] += norm(qlmiAve[qlmi_index]);
+                }
+                m_qliAve[qliAve_index] *= normalizationfactor[l_index];
+                m_qliAve[qliAve_index] = std::sqrt(m_qliAve[qliAve_index]);
             }
-            m_qliAve[i] *= normalizationfactor;
-            m_qliAve[i] = std::sqrt(m_qliAve[i]);
         });
 }
 
-float Steinhardt::normalizeSystem()
+std::vector<float> Steinhardt::normalizeSystem()
 {
-    float calc_norm(0);
-    const auto normalizationfactor = float(4.0 * M_PI / m_num_ms);
-    for (unsigned int k = 0; k < m_num_ms; ++k)
+    std::vector<float> system_norms(m_ls.size());
+    for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
     {
-        // Add the norm, which is the complex squared magnitude
-        calc_norm += norm(m_qlm[k]);
-    }
-    const float ql_system_norm = std::sqrt(calc_norm * normalizationfactor);
-
-    if (m_wl)
-    {
-        auto wigner3jvalues = getWigner3j(m_l);
-        float wl_system_norm = reduceWigner3j(m_qlm.get(), m_l, wigner3jvalues);
-
-        // The normalization factor of wl is calculated using qli, which is
-        // equivalent to calculate the normalization factor from qlmi
-        if (m_wl_normalize)
+        auto& qlm = m_qlm[l_index];
+        auto l = m_ls[l_index];
+        float calc_norm(0);
+        const auto normalizationfactor = static_cast<float>(4.0 * M_PI / m_num_ms[l_index]);
+        for (size_t k = 0; k < m_num_ms[l_index]; ++k)
         {
-            const float wl_normalization = std::sqrt(normalizationfactor) / ql_system_norm;
-            wl_system_norm *= wl_normalization * wl_normalization * wl_normalization;
+            // Add the norm, which is the complex squared magnitude
+            calc_norm += norm(qlm[k]);
         }
-        return wl_system_norm;
-    }
+        const float ql_system_norm = std::sqrt(calc_norm * normalizationfactor);
 
-    return ql_system_norm;
+        if (m_wl)
+        {
+            const auto wigner3j_values = getWigner3j(l);
+            float wl_system_norm = reduceWigner3j(qlm.get(), l, wigner3j_values);
+
+            // The normalization factor of wl is calculated using qli, which is
+            // equivalent to calculate the normalization factor from qlmi
+            if (m_wl_normalize)
+            {
+                const float wl_normalization = std::sqrt(normalizationfactor) / ql_system_norm;
+                wl_system_norm *= wl_normalization * wl_normalization * wl_normalization;
+            }
+            system_norms[l_index] = wl_system_norm;
+        }
+        else
+        {
+            system_norms[l_index] = ql_system_norm;
+        }
+    }
+    return system_norms;
 }
 
 void Steinhardt::aggregatewl(util::ManagedArray<float>& target,
-                             const util::ManagedArray<std::complex<float>>& source,
+                             const std::vector<util::ManagedArray<std::complex<float>>>& source,
                              const util::ManagedArray<float>& normalization_source) const
 {
-    auto wigner3jvalues = getWigner3j(m_l);
-    const auto normalizationfactor = float(4.0 * M_PI / m_num_ms);
     util::forLoopWrapper(0, m_Np, [&](size_t begin, size_t end) {
         for (size_t i = begin; i < end; ++i)
         {
-            target[i] = reduceWigner3j(&(source({static_cast<unsigned int>(i), 0})), m_l, wigner3jvalues);
-            if (m_wl_normalize)
+            const auto target_particle_index = target.getIndex({i, 0});
+            const auto norm_particle_index = normalization_source.getIndex({i, 0});
+            for (size_t l_index = 0; l_index < m_ls.size(); ++l_index)
             {
-                const float normalization = std::sqrt(normalizationfactor) / normalization_source[i];
-                target[i] *= normalization * normalization * normalization;
+                const auto l = m_ls[l_index];
+                const auto& source_l = source[l_index];
+
+                const auto normalizationfactor = static_cast<float>(4.0 * M_PI / m_num_ms[l_index]);
+                const auto wigner3j_values = getWigner3j(l);
+
+                target[target_particle_index + l_index]
+                    = reduceWigner3j(&(source_l({i, 0})), l, wigner3j_values);
+                if (m_wl_normalize)
+                {
+                    const float normalization = std::sqrt(normalizationfactor) /
+                        normalization_source[norm_particle_index + l_index];
+                    target[target_particle_index + l_index] *= normalization * normalization * normalization;
+                }
             }
         }
-    });
+        });
 }
 
 }; }; // end namespace freud::order