+++ ChangeLog.md
@@ -9,6 +9,7 @@ and this project adheres to
 ### Added
 * Added `out` option for the `wrap`, `unwrap`, `make_absolute`, and `make_fractional` methods of `Box`.
 * The Steinhardt and SolidLiquid classes expose the raw qlmi arrays.
+* The `Steinhardt` class supports computing order parameters for multiple `l`.
 
 ### Fixed
 * Fix broken arXiv links in bibliography.